+++ lib/main.dart
@@ -8,9 +8,9 @@ import 'package:receipt_wrangler_mobile/groups/nav/group/group_app_bar.dart';
 import 'package:receipt_wrangler_mobile/groups/nav/group/group_bottom_nav.dart';
 import 'package:receipt_wrangler_mobile/groups/nav/group_select/group_select_app_bar.dart';
 import 'package:receipt_wrangler_mobile/groups/nav/group_select/group_select_bottom_nav.dart';
-import 'package:receipt_wrangler_mobile/groups/screens/group-dashboards.dart';
-import 'package:receipt_wrangler_mobile/groups/screens/group-receipts-screen.dart';
-import 'package:receipt_wrangler_mobile/groups/screens/group-select.dart';
+import 'package:receipt_wrangler_mobile/groups/screens/group_dashboards.dart';
+import 'package:receipt_wrangler_mobile/groups/screens/group_receipts_screen.dart';
+import 'package:receipt_wrangler_mobile/groups/screens/group_select.dart';
 import 'package:receipt_wrangler_mobile/guards/auth-guard.dart';
 import 'package:receipt_wrangler_mobile/home/screens/home.dart';
 import 'package:receipt_wrangler_mobile/models/auth_model.dart';
